Splitting up ProjectionDaemon, renaming IProjectionDaemon API
