+++ docs/configuration/retries.md
@@ -12,7 +12,7 @@ Out of the box, Marten is using Polly for resiliency on most operations with thi
 // Default Polly setup
 var strategy = new ResiliencePipelineBuilder().AddRetry(new()
 {
-    ShouldHandle = new PredicateBuilder().Handle<NpgsqlException>().Handle<MartenCommandException>(),
+    ShouldHandle = new PredicateBuilder().Handle<NpgsqlException>().Handle<MartenCommandException>().Handle<EventLoaderException>(),
     MaxRetryAttempts = 3,
     Delay = TimeSpan.FromMilliseconds(50),
     BackoffType = DelayBackoffType.Exponential
