+++ docs/events/projections/async-daemon.md
@@ -328,7 +328,7 @@ public static async Task UseAsyncDaemon(IDocumentStore store, CancellationToken 
     using var daemon = await store.BuildProjectionDaemonAsync();
 
     // Fire up everything!
-    await daemon.StartAllShards();
+    await daemon.StartAllAsync();
 
     // or instead, rebuild a single projection
     await daemon.RebuildProjection("a projection name", 5.Minutes(), cancellation);
@@ -344,7 +344,7 @@ public static async Task UseAsyncDaemon(IDocumentStore store, CancellationToken 
     await daemon.StartShard("shard name", cancellation);
 
     // Or change your mind and stop the shard you just started
-    await daemon.StopShard("shard name");
+    await daemon.StopAsync("shard name");
 
     // No, shut them all down!
     await daemon.StopAllAsync();
