+++ src/Marten.AsyncDaemon.Testing/Bugs/Bug_2073_tenancy_problems.cs
@@ -52,7 +52,7 @@ public class Bug_2073_tenancy_problems
         var store = host.Services.GetRequiredService<IDocumentStore>();
         await store.Advanced.Clean.CompletelyRemoveAllAsync();
         var daemon = (ProjectionDaemon)(await store.BuildProjectionDaemonAsync());
-        await daemon.StartAllShards();
+        await daemon.StartAllAsync();
 
         await using (var session = store.LightweightSession("tenant1"))
         {
@@ -62,7 +62,7 @@ public class Bug_2073_tenancy_problems
 
         await daemon.Tracker.WaitForShardState("Document:All", 1);
 
-        daemon.CurrentShards().Single()
+        daemon.CurrentAgents().Single()
             .Status.ShouldBe(AgentStatus.Running);
 
         await daemon.StopAllAsync();
