+++ src/Marten.AsyncDaemon.Testing/DocumentTrackingByIdentity/EventProjectionWithCreate_follow_up_operations.cs
@@ -55,7 +55,7 @@ public class EventProjectionWithCreate_follow_up_operations: DaemonContext
 
         var daemon = await StartDaemon();
 
-        await daemon.StartDaemonAsync();
+        await daemon.StartHighWaterDetectionAsync();
 
         await daemon.Tracker.WaitForShardState($"{nameof(EntityProjection)}:All", 2);
