+++ src/Marten.AsyncDaemon.Testing/HighWaterAgentTests.cs
@@ -54,7 +54,7 @@ public class HighWaterAgentTests: DaemonContext
         await agent.StopAllAsync();
 
         using var agent2 = await theStore.BuildProjectionDaemonAsync();
-        await agent2.StartDaemonAsync();
+        await agent2.StartHighWaterDetectionAsync();
         await agent2.Tracker.WaitForHighWaterMark(NumberOfEvents, 15.Seconds());
 
     }
